requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
Django==5.1.6
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-filter==24.3
django-js-asset==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.1
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
